.history/Restaurant_dailysale/requirements_20260305114418.txt
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
pytesseract==0.3.13
